+++ ci/integration/ubuntu-entrypoint.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+
+# Start the docker daemon in the containr
+service docker start
+
+exec "$@"